@@ -18,27 +18,27 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <=3.2.4"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  expandable: ^4.1.4
-  intl: ^0.17.0
-  flutter_launcher_icons: ^0.8.1
-  flutter_colorpicker: ^0.3.5
-  flutter_iconpicker: ^2.2.2
-  firebase_core: ^1.0.0
+  expandable: ^5.0.1
+  intl: ^0.18.1
+  flutter_launcher_icons: ^0.13.1
+  flutter_colorpicker: ^1.0.3
+  flutter_iconpicker: ^3.2.4
+  firebase_core: ^2.24.2
   firebase: ^9.0.0
-  cloud_firestore: ^1.0.0
-  firebase_auth: ^1.0.0
-  google_sign_in: ^5.0.0
-  mask_text_input_formatter: ^1.2.1
-  flutter_masked_text: ^0.8.0
+  cloud_firestore: ^4.14.0
+  firebase_auth: ^4.16.0
+  google_sign_in: ^6.2.1
+  mask_text_input_formatter: ^2.7.0
+  flutter_masked_text2: ^0.9.1
   shared_preferences: ^2.0.3
-  syncfusion_flutter_charts: ^18.4.46
+  syncfusion_flutter_charts: ^24.1.45+2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
